DIV-3113 - moved capture case and pin logic into loadMiniPetition
