+++ steps/authenticated/Authenticated.step.js
@@ -2,7 +2,6 @@ const { Redirect } = require('@hmcts/one-per-page');
 const { redirectTo } = require('@hmcts/one-per-page/flow');
 const idam = require('services/idam');
 const config = require('config');
-const middleware = require('middleware/authenticated');
 
 class Authenticated extends Redirect {
   static get path() {
@@ -16,8 +15,7 @@ class Authenticated extends Redirect {
   get middleware() {
     return [
       ...super.middleware,
-      idam.landingPage(),
-      middleware.captureCaseAndPin
+      idam.landingPage()
     ];
   }
 }
