+++ test/unit/steps/Authenicated.step.test.js
@@ -1,7 +1,6 @@
 const modulePath = 'steps/authenticated/Authenticated.step';
 
 const Authenticated = require(modulePath);
-const authenticatedMiddleware = require('middleware/authenticated');
 const Respond = require('steps/respond/Respond.step');
 const idam = require('services/idam');
 const { middleware, redirect, sinon } = require('@hmcts/one-per-page-test-suite');
@@ -10,20 +9,14 @@ describe(modulePath, () => {
   beforeEach(() => {
     sinon.stub(idam, 'landingPage')
       .returns(middleware.nextMock);
-    sinon.stub(authenticatedMiddleware, 'captureCaseAndPin')
-      .callsFake(middleware.nextMock);
   });
 
   afterEach(() => {
     idam.landingPage.restore();
-    authenticatedMiddleware.captureCaseAndPin.restore();
   });
 
-  it('has idam.landingPage and captureCaseAndPin middleware', () => {
-    return middleware.hasMiddleware(Authenticated, [
-      idam.landingPage(),
-      authenticatedMiddleware.captureCaseAndPin
-    ]);
+  it('has idam.landingPage middleware', () => {
+    return middleware.hasMiddleware(Authenticated, [idam.landingPage()]);
   });
 
   context('navigation', () => {
