+++ test/unit/steps/respond.test.js
@@ -5,22 +5,18 @@ const ReviewApplication = require('steps/review-application/ReviewApplication.st
 const idam = require('services/idam');
 const { middleware, interstitial, sinon, content } = require('@hmcts/one-per-page-test-suite');
 const getSteps = require('steps');
-const authenticated = require('middleware/authenticated');
 const loadMiniPetition = require('middleware/loadMiniPetition');
 
 describe(modulePath, () => {
   beforeEach(() => {
     sinon.stub(idam, 'protect')
       .returns(middleware.nextMock);
-    sinon.stub(authenticated, 'captureCaseAndPin')
-      .callsFake(middleware.nextMock);
     sinon.stub(loadMiniPetition, 'loadMiniPetition')
       .callsFake(middleware.nextMock);
   });
 
   afterEach(() => {
     idam.protect.restore();
-    authenticated.captureCaseAndPin.restore();
     loadMiniPetition.loadMiniPetition.restore();
   });
